requirements.txt
discord.py
dropbox